refactor: isolate daemon artifact downloads
